+++ Sources/Sentry/SentryHub.m
@@ -760,7 +760,7 @@ NS_ASSUME_NONNULL_BEGIN
     for (SentryEnvelopeItem *item in items) {
         if ([item.header.type isEqualToString:SentryEnvelopeItemTypeEvent]) {
             // If there is no level the default is error
-            NSDictionary *eventJson =
+            NSDictionary *_Nullable eventJson =
                 [SentrySerialization deserializeDictionaryFromJsonData:item.data];
             if (eventJson == nil) {
                 return NO;