+++ Tests/SentryTests/Integrations/Feedback/SentryFeedbackTests.swift
@@ -1,6 +1,7 @@
 import Foundation
 #if os(iOS) && !SENTRY_NO_UIKIT
-@testable import Sentry
+@_spi(Private) @testable import Sentry
+@_spi(Private) import SentryTestUtils
 import XCTest
 
 class SentryFeedbackTests: XCTestCase {
@@ -178,15 +179,15 @@ class SentryFeedbackTests: XCTestCase {
             func testCaseDescription() -> String {
                 "(config: (requiresName: \(input.config.requiresName), requiresEmail: \(input.config.requiresEmail), nameInput: \(input.config.nameInput == nil ? "nil" : "\"\(input.config.nameInput!)\""), emailInput: \(input.config.emailInput == nil ? "nil" : "\"\(input.config.emailInput!)\""), messageInput: \(input.config.messageInput == nil ? "nil" : "\"\(input.config.messageInput!)\""), includeScreenshot: \(input.config.includeScreenshot)), expectedSubmitButtonAccessibilityHint: \(input.expectedSubmitButtonAccessibilityHint)"
             }
-            SentryLog.withoutLogs {
-                switch viewModel.validate() {
-                case .success(let hint):
-                    XCTAssert(input.shouldValidate)
-                    XCTAssertEqual(hint, input.expectedSubmitButtonAccessibilityHint, testCaseDescription())
-                case .failure(let error):
-                    XCTAssertFalse(input.shouldValidate, error.description + "; " + testCaseDescription())
-                }
+
+            switch viewModel.validate() {
+            case .success(let hint):
+                XCTAssert(input.shouldValidate)
+                XCTAssertEqual(hint, input.expectedSubmitButtonAccessibilityHint, testCaseDescription())
+            case .failure(let error):
+                XCTAssertFalse(input.shouldValidate, error.description + "; " + testCaseDescription())
             }
+
         }
     }
 }