+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Fixes
+
+- Add missing context for watchdog termination events (#5242)
+
+## 8.52.0-beta
+
 ### Features
 
 - XCFrameworks are now signed (#5271)
@@ -9,6 +15,7 @@
 ### Improvements
 
 - Slightly reduce performance impact by removing unnecessary lock in SentryLog.configure (#5297)
+- Redact React Native text and images by default without the RN SDK (#5302)
 
 ## 8.51.1